Update to allow for sizing a batch
This update will allow for the ability to have a batchsizeupperlimit - this value (in the T3KAI.json   "t3kaibatchsizeupperlimit": 5000,) will allow the user to set the batch size to be sent to T3KAI - if the user sets this value to 0 all the items selected (in the T3KAI Analyze Type field) will be sent as one batch - if this value is set to a number batch sizes of that number will be sent to T3KAI at a time.  i.e. if there are 50,000 items selected in the interface and the t3kaibatchsizeupperlimit value is set to 5,000 then 10 batches of 5,000 items will be sent to CORE for analysis.