Connect the UI to the T3K API
Move the application.rb and related code to relative to the in-Nuix scripts

Call the application.rb code from the in-Nuix code

Modify application.rb to use the progress dialog

Move the settings to the top of the scripts folder and the id-cache to the scripts folder.

Make the to_s method on results nicer.